change Field to FieldPreview & Add Field, District
