+++ src/main/kotlin/jp/trap/conqest/game/Game.kt
@@ -6,7 +6,7 @@ import org.bukkit.entity.Player
 import org.bukkit.plugin.Plugin
 import java.util.*
 
-class Game(val plugin: Plugin) {
+class Game(val plugin: Plugin, val field: Field) {
 
     private var state: GameState = GameState.BeforeGame(this)
     private val players: MutableList<Player> = mutableListOf()
