+++ src/main/kotlin/jp/trap/conqest/game/GameManager.kt
@@ -6,14 +6,19 @@ import org.bukkit.plugin.Plugin
 class GameManager(val plugin: Plugin) {
     private val games: MutableList<Game> = mutableListOf()
     private val fields: MutableList<Field> = mutableListOf()
-    fun requestOpeningGame(): Game {
+    fun requestOpeningGame(): Game? {
         return games.firstOrNull { game -> game.getStateType() == GameStates.MATCHING }
-            ?: games.firstOrNull { game -> game.getStateType() == GameStates.BEFORE_GAME }
-            ?: run {
-                val newGame = Game(plugin)
-                games.add(newGame)
-                newGame
+            ?: games.firstOrNull { game -> game.getStateType() == GameStates.BEFORE_GAME } ?: run {
+                requestOpeningField()?.let { field ->
+                    val newGame = Game(plugin, field)
+                    games += newGame
+                    newGame
+                }
             }
+    }
+
+    private fun requestOpeningField(): Field? {
+        return fields.firstOrNull { field -> games.all { game -> game.field != field } }
     }
 
     fun getGame(player: Player): Game? {
