+++ src/main/kotlin/jp/trap/conqest/game/Team.kt
@@ -1,0 +1,4 @@
+package jp.trap.conqest.game
+
+class Team {
+}
